+++ src/index.ts
@@ -1,5 +1,13 @@
+/**
+ * Style Export
+ */
+
 import './styles/tailwind.css'
 import './index.scss'
+
+/**
+ * Component Export
+ */
 
 export * from './components/Button'
 export * from './components/Checkbox'
@@ -17,3 +25,10 @@ export * from './components/Switch'
 export * from './components/Table'
 export * from './components/Tag'
 export * from './components/Toast'
+
+/**
+ * Util Export
+ */
+
+export { UUI } from './utils/uui'
+export type { UUIClassComponentProps, UUIFunctionComponentProps, UUIComponentCustomizeProps } from './utils/uui'